split implementation of store in several files
